+++ packages/payment-detection/src/any/any-to-erc20-proxy-info-retriever.ts
@@ -1,0 +1,20 @@
+import ConversionInfoRetriever from './any-to-any-proxy-info-retriever';
+import { ethers } from 'ethers';
+
+/**
+ * Retrieves a list of payment events from a payment reference, a destination address and a proxy contract
+ */
+export default class AnyToErc20InfoRetriever extends ConversionInfoRetriever {
+  protected getFeeFilter(): ethers.providers.Filter {
+    // Create a filter to find all the Fee Transfer logs with the payment reference
+    const feeFilter = this.contractConversionProxy.filters.TransferWithReferenceAndFee(
+      null,
+      null,
+      null,
+      '0x' + this.paymentReference,
+    ) as ethers.providers.Filter;
+    feeFilter.fromBlock = this.conversionProxyCreationBlockNumber;
+    feeFilter.toBlock = 'latest';
+    return feeFilter;
+  }
+}
